AGS.Native: fixed several cases of incorrect sprintf argument

+++ Engine/debug/debug.cpp
@@ -259,7 +259,7 @@ bool send_message_to_editor(const char *msg, const char *errorMsg)
     char messageToSend[STD_BUFFER_SIZE];
     sprintf(messageToSend, "<?xml version=\"1.0\" encoding=\"Windows-1252\"?><Debugger Command=\"%s\">", msg);
 #ifdef WINDOWS_VERSION
-    sprintf(&messageToSend[strlen(messageToSend)], "  <EngineWindow>%d</EngineWindow> ", win_get_window());
+    sprintf(&messageToSend[strlen(messageToSend)], "  <EngineWindow>%d</EngineWindow> ", (int)win_get_window());
 #endif
     sprintf(&messageToSend[strlen(messageToSend)], "  <ScriptState><![CDATA[%s]]></ScriptState> ", callStack.GetCStr());
     if (errorMsg != NULL)
